Move DRAND_MAINNET2_PUBKEY to drand_common
