+++ contracts/nois-drand/src/lib.rs
@@ -1,6 +1,5 @@
 mod attributes;
 mod bots;
-mod drand;
 
 pub mod contract;
 pub mod error;
